added further test cases

+++ src/main/java/com/digipay/rest/dtos/TransactionDto.java
@@ -2,7 +2,10 @@ package com.digipay.rest.dtos;
 
 import com.digipay.entities.Transaction;
 import com.digipay.entities.TransactionStatus;
+import lombok.Builder;
 import lombok.Data;
+
+import java.math.BigDecimal;
 
 @Data
 public class TransactionDto {
@@ -10,13 +13,13 @@ public class TransactionDto {
     private String transactionId;
     private String senderId;
     private String receiverId;
-    private Double amount;
+    private BigDecimal amount;
     private TransactionStatus status;
     private String remark;
 
     public static TransactionDto from(Transaction transaction) {
         TransactionDto dto = new TransactionDto();
-        dto.amount = transaction.getAmount().doubleValue();
+        dto.amount = transaction.getAmount();
         dto.senderId = transaction.getSenderId();
         dto.receiverId = transaction.getReceiverId();
         dto.transactionId = transaction.getId();
