+++ src/main/java/com/digipay/rest/dtos/UserDto.java
@@ -4,19 +4,21 @@ import com.digipay.entities.User;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class UserDto {
 
     private String id;
     private String name;
     @JsonProperty("balance")
-    private Double balance;
+    private BigDecimal balance;
 
     public static UserDto from(User user) {
         UserDto dto = new UserDto();
         dto.id = user.getId();
         dto.name = user.getName();
-        dto.balance = user.getBalance().doubleValue();
+        dto.balance = user.getBalance();
         return dto;
     }
 }
